Update README.md
This project was created as part of a final course in the GIS course. 
the site can add layers to the regular map, Each layer provides a different solution for finding accessibility within the cities, such as where there are public transport stations and where they pass etc, you can also mark the current location on the map, and you can move the map using coordinates That the user enters.